Add tests and move visitor to best_practices
