+++ tests/fixtures/noqa.py
@@ -327,7 +327,7 @@ print(one == 'a' or one == 'b')  # noqa: WPS514
 file_obj = open('filaname.py')  # noqa: WPS515
 print(type(file_obj) == int)  # noqa: WPS516
 
-print(*[])  # noqa: WPS517
+print(*[], **{'@': 1})  # noqa: WPS517, WPS445
 
 for range_len in range(len(file_obj)):  # noqa: WPS518
     print(range_len)
